feat: add the scoring system
Add 1 point on each click on the image that hasn't been clicked before.
If the image has already been clicked before, check if the current score
should be the next high score and make it so if it is. After, reset the
current score back to 0 and reset the clickedImages state for a new game.